requirements.txt
absl-py==1.2.0
appdirs==1.4.4
argcomplete==2.0.0
audioread==3.0.0
beautifulsoup4==4.11.1
Bottleneck==1.3.5
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
ci-sdr==0.0.2
click==8.1.3
ConfigArgParse==1.5.3
ctc-segmentation==1.7.3
cycler==0.11.0
Cython==0.29.32
decorator==5.1.1
Distance==0.1.3
docker-pycreds==0.4.0
editdistance==0.5.2
einops==0.4.1
espnet==0.10.0
espnet-tts-frontend==0.0.3
filelock==3.8.0
Flask==2.2.2
g2p-en==2.1.0
gdown==4.5.1
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.49.0
h5py==3.7.0
humanfriendly==10.0
idna==3.4
importlib-metadata==4.12.0
inflect==6.0.0
itsdangerous==2.1.2
jaconv==0.3
Jinja2==3.1.2
joblib==1.1.0
kaldiio==2.17.2
kiwisolver==1.4.4
librosa==0.9.2
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.1.0
nara-wpe==0.0.8
nltk==3.7
numba==0.56.2
numpy==1.23.3
oauthlib==3.2.1
packaging==21.3
parallel-wavegan==0.4.8
pathtools==0.1.2
Pillow==9.2.0
pooch==1.6.0
promise==2.3
protobuf==3.19.5
psutil==5.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pyopenjtalk==0.1.3
pyparsing==3.0.9
pypinyin==0.47.1
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-wpe==0.0.1
pyworld==0.3.0
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
sentencepiece==0.1.86
sentry-sdk==1.9.8
setproctitle==1.3.2
shortuuid==1.0.9
simpleaudio==1.0.4
six==1.16.0
smmap==5.0.0
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
threadpoolctl==3.1.0
toml==0.10.2
torch==1.13.0.dev20220913
torch-complex==0.4.3
torchaudio==0.13.0.dev20220915
torchvision==0.14.0.dev20220915
tqdm==4.64.1
typeguard==2.13.3
typing_extensions==4.3.0
Unidecode==1.3.4
urllib3==1.26.12
wandb==0.13.3
Werkzeug==2.2.2
xmltodict==0.13.0
yq==3.1.0
zipp==3.8.1
